fix(handlers): :bug: Handle IPv6 address for development server URL (#19)
This commit fixes an issue where the _getDevServerUrl function displayed [::1] instead of localhost for the development server address when using IPv6.  Updates the function to check if the host is [::1] (IPv6 loopback address) and replaces it with localhost in the returned URL. Improves user experience by displaying a more familiar address during development.